Don't pass additional requirements to the resolver
Downgrading is actually necessary sometimes. For example, when upgrading
manageiq (https://github.com/ManageIQ/manageiq) lockfile generated with
older versions of bundler to use the new format with separated GEM
sections, the result of the secure resolution results in mime-types
getting now picked up from the custom `https://rubygems.manageiq.org`
source, which only has mime-types 3.0.0 (as opposed to the locked 3.3.1
version).

Furthermore, these extra requirements seems to cause infinite resolution
loops sometimes.

Finally, none of the specs covering downgrades where not actually wanted
fail when removing this, so it seems that the resolver itself is already
handling this in the most reasonable way.

So, I'm tentatively removing this code.